+++ microsite/docusaurus.config.js
@@ -16,6 +16,11 @@
 
 // @ts-check
 
+/** @type{import('prism-react-renderer').PrismTheme} **/
+// @ts-ignore
+const prismTheme = require('prism-react-renderer/themes/vsDark');
+prismTheme.plain.backgroundColor = '#232323';
+
 /** @type {import('@docusaurus/types').Config} */
 module.exports = {
   title: 'Backstage Software Catalog and Developer Platform',
@@ -63,6 +68,12 @@ module.exports = {
       },
     ],
   ],
+  markdown: {
+    preprocessor({ filePath, fileContent }) {
+      // Replace all HTML comments with emtpy strings as these are not supported by MDXv2.
+      return fileContent.replace(/<!--.*?-->/gs, '');
+    },
+  },
   webpack: {
     jsLoader: isServer => ({
       loader: require.resolve('swc-loader'),
@@ -119,6 +130,10 @@ module.exports = {
           {
             from: '/docs/features/search/search-overview',
             to: '/docs/features/search/',
+          },
+          {
+            from: '/docs/getting-started/running-backstage-locally',
+            to: '/docs/getting-started/',
           },
         ],
       },
@@ -266,10 +281,11 @@ module.exports = {
         searchParameters: {},
       },
       prism: {
+        theme: prismTheme,
         magicComments: [
           // Extend the default highlight class name
           {
-            className: 'theme-code-block-highlighted-line',
+            className: 'code-block-highlight-line',
             line: 'highlight-next-line',
             block: { start: 'highlight-start', end: 'highlight-end' },
           },