+++ plugins/scaffolder-backend/src/scaffolder/tasks/NunjucksWorkflowRunner.ts
@@ -26,11 +26,10 @@ import fs from 'fs-extra';
 import path from 'path';
 import nunjucks from 'nunjucks';
 import { JsonObject, JsonValue } from '@backstage/types';
-import { InputError } from '@backstage/errors';
+import { InputError, NotAllowedError } from '@backstage/errors';
 import { PassThrough } from 'stream';
 import { generateExampleOutput, isTruthy } from './helper';
 import { validate as validateJsonSchema } from 'jsonschema';
-import { parseRepoUrl } from '../actions/builtin/publish/util';
 import { TemplateActionRegistry } from '../actions';
 import {
   TemplateFilter,
@@ -43,9 +42,19 @@ import {
   TaskSpecV1beta3,
   TaskStep,
 } from '@backstage/plugin-scaffolder-common';
+
 import { TemplateAction } from '@backstage/plugin-scaffolder-node';
+import { createConditionAuthorizer } from '@backstage/plugin-permission-node';
 import { UserEntity } from '@backstage/catalog-model';
 import { createCounterMetric, createHistogramMetric } from '../../util/metrics';
+import { createDefaultFilters } from '../../lib/templating/filters';
+import {
+  AuthorizeResult,
+  PermissionEvaluator,
+  PolicyDecision,
+} from '@backstage/plugin-permission-common';
+import { scaffolderActionRules } from '../../service/rules';
+import { actionExecutePermission } from '@backstage/plugin-scaffolder-common/alpha';
 
 type NunjucksWorkflowRunnerOptions = {
   workingDirectory: string;
@@ -54,6 +63,7 @@ type NunjucksWorkflowRunnerOptions = {
   logger: winston.Logger;
   additionalTemplateFilters?: Record<string, TemplateFilter>;
   additionalTemplateGlobals?: Record<string, TemplateGlobal>;
+  permissions?: PermissionEvaluator;
 };
 
 type TemplateContext = {
@@ -102,8 +112,17 @@ const createStepLogger = ({
   return { taskLogger, streamLogger };
 };
 
+const isActionAuthorized = createConditionAuthorizer(
+  Object.values(scaffolderActionRules),
+);
+
 export class NunjucksWorkflowRunner implements WorkflowRunner {
-  constructor(private readonly options: NunjucksWorkflowRunnerOptions) {}
+  private readonly defaultTemplateFilters: Record<string, TemplateFilter>;
+  constructor(private readonly options: NunjucksWorkflowRunnerOptions) {
+    this.defaultTemplateFilters = createDefaultFilters({
+      integrations: this.options.integrations,
+    });
+  }
 
   private readonly tracker = scaffoldingTracker();
 
@@ -193,6 +212,7 @@ export class NunjucksWorkflowRunner implements WorkflowRunner {
     renderTemplate: (template: string, values: unknown) => string,
     taskTrack: TaskTrackType,
     workspacePath: string,
+    decision: PolicyDecision,
   ) {
     const stepTrack = await this.tracker.stepStart(task, step);
 
@@ -276,6 +296,18 @@ export class NunjucksWorkflowRunner implements WorkflowRunner {
         }
       }
 
+      if (!isActionAuthorized(decision, { action: action.id, input })) {
+        throw new NotAllowedError(
+          `Unauthorized action: ${
+            action.id
+          }. The action is not allowed. Input: ${JSON.stringify(
+            input,
+            null,
+            2,
+          )}`,
+        );
+      }
+
       const tmpDirs = new Array<string>();
       const stepOutput: { [outputName: string]: JsonValue } = {};
 
@@ -329,22 +361,15 @@ export class NunjucksWorkflowRunner implements WorkflowRunner {
       await task.getWorkspaceName(),
     );
 
-    const {
-      additionalTemplateFilters,
-      additionalTemplateGlobals,
-      integrations,
-    } = this.options;
+    const { additionalTemplateFilters, additionalTemplateGlobals } =
+      this.options;
 
     const renderTemplate = await SecureTemplater.loadRenderer({
-      // TODO(blam): let's work out how we can deprecate this.
-      // We shouldn't really need to be exposing these now we can deal with
-      // objects in the params block.
-      // Maybe we can expose a new RepoUrlPicker with secrets for V3 that provides an object already.
-      parseRepoUrl(url: string) {
-        return parseRepoUrl(url, integrations);
+      templateFilters: {
+        ...this.defaultTemplateFilters,
+        ...additionalTemplateFilters,
       },
-      additionalTemplateFilters,
-      additionalTemplateGlobals,
+      templateGlobals: additionalTemplateGlobals,
     });
 
     try {
@@ -357,6 +382,14 @@ export class NunjucksWorkflowRunner implements WorkflowRunner {
         user: task.spec.user,
       };
 
+      const [decision]: PolicyDecision[] =
+        this.options.permissions && task.spec.steps.length
+          ? await this.options.permissions.authorizeConditional(
+              [{ permission: actionExecutePermission }],
+              { token: task.secrets?.backstageToken },
+            )
+          : [{ result: AuthorizeResult.ALLOW }];
+
       for (const step of task.spec.steps) {
         await this.executeStep(
           task,
@@ -365,6 +398,7 @@ export class NunjucksWorkflowRunner implements WorkflowRunner {
           renderTemplate,
           taskTrack,
           workspacePath,
+          decision,
         );
       }