+++ plugins/jenkins-common/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @backstage/plugin-jenkins-common
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.7.5
+  - @backstage/plugin-catalog-common@1.0.13
+
+## 0.1.15-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-common@1.0.13-next.1
+  - @backstage/plugin-permission-common@0.7.5-next.0
+
+## 0.1.15-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.7.5-next.0
+  - @backstage/plugin-catalog-common@1.0.13-next.0
+
 ## 0.1.14
 
 ### Patch Changes